+++ src/core/valuemapmodelbase.cpp
@@ -55,7 +55,7 @@ void ValueMapModelBase::setMap( const QVariant &map )
     const QVariantMap valueMap = map.toMap();
     if ( !valueMap.empty() )
     {
-      beginInsertRows( QModelIndex(), 0, valueMap.size() - 1 );
+      beginInsertRows( QModelIndex(), 0, static_cast<int>( valueMap.size() ) - 1 );
 
       QMapIterator<QString, QVariant> i( valueMap );
       while ( i.hasNext() )
@@ -77,7 +77,7 @@ void ValueMapModelBase::setMap( const QVariant &map )
 int ValueMapModelBase::rowCount( const QModelIndex &parent ) const
 {
   Q_UNUSED( parent )
-  return mMap.size();
+  return static_cast<int>( mMap.size() );
 }
 
 QVariant ValueMapModelBase::data( const QModelIndex &index, int role ) const
